Python Project1 (Very, very simple text editor)
This is my first project in GitHub or anywhere! This project I built out of my new knowledge of python file management, and it consists of a very simple application that writes lines to a file named "yourfile.txt". Controls are clear for clearing terminal, ~ for ending the program(and creating the txt file) and erase to delete the previously saved text line.