Project 3 finality
Project3 is somewhat of a failure. It doesn't work well! I've assumed a very large success rate for dogs, yet cats it has trouble with. I'm assuming it's due to weird training (epoch = 1000, lack of images, weird ordering between Conv2D and MaxPooling, etc.) Well, this has been a great starting point, but on to PyTorch! (and hopefully CUDA training)